First Steps
I have decided on a minimalist design for the website and will most likely stick to that design style for the time being.

With this new commit I have finished the navigation bar and added a password system. It is in place just so there is some kind of wall preventing a player from accessing this website. Will be beefing it up in the future but mostly there for proof of concept.

Also, please enjoy that place holder image.

Next commit I do will be on the Development Path. I will be giving an in order list of the things to come. They will come in parts. Part 1 will entail everything that will be completed before the start of the campaign. Part 2 will have everything I will be making for the next 3 to 4 months. Part 3 will have everything I will be making in the coming year or 2.